+++ server/package.json
@@ -7,7 +7,10 @@
     "typescript": "^4.9.3"
   },
   "dependencies": {
+    "@types/body-parser": "^1.19.2",
+    "body-parser": "^1.20.1",
     "express": "^4.18.2",
-    "mongoose": "^6.7.5"
+    "mongoose": "^6.7.5",
+    "ts-node": "^10.9.1"
   }
 }
